requirements.txt
python-dotenv
discord.py
wit